Add button to enable mask layer selection
Introduces a 'Provide Mask Layer' button that, when clicked, reveals the mask layer combobox and label. The combobox and label are initially hidden, and are hidden again if the user selects 'None'. This improves the UI by making mask selection optional and less cluttered by default. Also adds a tooltip to the combobox for user guidance.