Filter selection by harmonic value
Added filtering to ensure only rows matching the current harmonic value are included in the selection.
This fixes a bug where changing the selection choice in the selection tab to None and back to MANUAL SELECTION #1 would raise an AttributeError due to trying to assign arrays of different sizes to `.color_indices`.