Refactor escape handling and improve drag logic
Replaced the local _on_escape method by the one in biaplotter and updated component selection and drag event handlers to call it from there. Enhanced _on_press method to deactivate zoom or pan after clicking (not covered by biaplotter _on_escape method) when dragging component text labels or dots. _on_press functionality is only allowed if the components_tab is the active one to avoid triggering it when accidently clicking on the spot where there is a component, but the dot or text label artists are hidden.